+++ projects/lux/src/lib/autocomplete/autocomplete.component.ts
@@ -325,14 +325,14 @@ export class AutocompleteComponent
     const searchText = (text || '').toLowerCase();
     if (this.dataSource) {
       const ds = (this.dataSource || [])
-        .filter((it) => it.label.toLowerCase().includes(searchText))
+        .filter((it) => ignoreAccentsInclude(it.label, searchText))
         .sort((a, b) => a.label.localeCompare(b.label));
       return of(decorateDataSource(ds, searchText));
     } else if (this.instance && this.populateFunction) {
       return this.populateFunction(this.instance, searchText).pipe(
         first(),
         map((ds) => {
-          ds.filter((it) => it.label.toLowerCase().includes(searchText)).sort(
+          ds.filter((it) => ignoreAccentsInclude(it.label, searchText)).sort(
             (a, b) => a.label.localeCompare(b.label)
           );
           return decorateDataSource(ds, searchText);
@@ -344,6 +344,17 @@ export class AutocompleteComponent
   }
 }
 
+/** Returns true if text includes substring. No accents considered. Ignorecase */
+const ignoreAccentsInclude = (text: string, substring: string): boolean => {
+  text = normalizedString(text).toLowerCase();
+  substring = normalizedString(substring).toLowerCase();
+  return text.includes(substring);
+};
+
+/** Returns a normalized string with no accents: used for comparison */
+const normalizedString = (a: string): string =>
+  (a || '').normalize('NFD').replace(/[\u0300-\u036f]/g, '');
+
 const decorateDataSource = (
   dataSource: DataSource<any, string>,
   subString: string
@@ -353,7 +364,9 @@ const decorateItem = (
   item: DataSourceItem<any, string>,
   tx: string
 ): DecoratedDataSourceItem => {
-  const index = item.label.toLowerCase().indexOf(tx.toLowerCase());
+  const index = normalizedString(item.label)
+    .toLowerCase()
+    .indexOf(normalizedString(tx).toLowerCase());
   const labelPrefix = index === -1 ? item.label : item.label.substr(0, index);
   const labelMatch = index === -1 ? '' : item.label.substr(index, tx.length);
   const labelPostfix = index === -1 ? '' : item.label.substr(index + tx.length);