+++ projects/lux/Changelog.md
@@ -1,5 +1,12 @@
 # Lux Changelog
 
+## Version 0.7.2
+
+2021.06.27
+
+- Ignore accents on search: autocomplete
+- Limit max & min dates to year 2100 and 1900 by default.
+
 ## Version 0.7.1
 
 2021.06.26